backend/requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
selenium>=4.15.0

# Database
psycopg2-binary>=2.9.9

# OpenAI
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
pytz>=2024.1

# Rate Limiting
ratelimit>=2.2.1

# Email (optional for notifications)
# Already in standard library: smtplib, email
